Zadanie 5: Sumowanie ilości produktów w koszyku

+++ src/main/java/pl/wsei/store/service/BasketService.java
@@ -13,12 +13,24 @@ public class BasketService {
 
     public void buyItem(String item, Integer quantity) {
         EntityManager em = emf.createEntityManager();
+
         try {
             em.getTransaction().begin();
 
-            for (int i = 0; i < quantity; i++) {
+            Basket existingBasket = em.createQuery(
+                "SELECT b FROM Basket b WHERE b.item = :item", Basket.class)
+                .setParameter("item", item)
+                .getResultStream()
+                .findFirst()
+                .orElse(null);
+
+            if (existingBasket != null) {
+                existingBasket.setQuantity(existingBasket.getQuantity() + quantity);
+                em.merge(existingBasket);
+            } else {
                 Basket basket = new Basket();
                 basket.setItem(item);
+                basket.setQuantity(quantity);
                 em.persist(basket);
             }
